# schema/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Infrahub",
        "version": "latest"
    },
    "paths": {
...
        "/api/schema": {
            "get": {
                "summary": "Get Schema",
                "operationId": "get_schema_api_schema_get",
                "security": [
                    {
                        "HTTPBearer": []
                    },
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "namespaces",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Namespaces"
                        }
                    },
                    {
                        "name": "branch",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Name of the branch to use for the query",
                            "title": "Branch"
                        },
                        "description": "Name of the branch to use for the query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SchemaReadAPI"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/schema/{schema_kind}": {
            "get": {
                "summary": "Get Schema By Kind",
                "operationId": "get_schema_by_kind_api_schema__schema_kind__get",
                "security": [
                    {
                        "HTTPBearer": []
                    },
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "schema_kind",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Schema Kind"
                        }
                    },
                    {
                        "name": "branch",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Name of the branch to use for the query",
                            "title": "Branch"
                        },
                        "description": "Name of the branch to use for the query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/APIProfileSchema"
                                        },
                                        {
                                            "$ref": "#/components/schemas/APINodeSchema"
                                        },
                                        {
                                            "$ref": "#/components/schemas/APIGenericSchema"
                                        },
                                        {
                                            "$ref": "#/components/schemas/APITemplateSchema"
                                        }
                                    ],
                                    "title": "Response Get Schema By Kind Api Schema  Schema Kind  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "APIProfileSchema": {
                "properties": {
                    "id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Id",
                        "description": "The ID of the node",
                        "update": "not_applicable"
                    },
                    "state": {
                        "$ref": "#/components/schemas/HashableModelState",
                        "description": "Expected state of the node/generic after loading the schema",
                        "default": "present",
                        "update": "not_applicable"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 32,
                        "minLength": 2,
                        "pattern": "^[A-Z][a-zA-Z0-9]+$",
                        "title": "Name",
                        "description": "Node name, must be unique within a namespace and must start with an uppercase letter.",
                        "update": "migration_required"
                    },
                    "namespace": {
                        "type": "string",
                        "maxLength": 64,
                        "minLength": 3,
                        "pattern": "^[A-Z][a-z0-9]+$",
                        "title": "Namespace",
                        "description": "Node Namespace, Namespaces are used to organize models into logical groups and to prevent name collisions.",
                        "update": "migration_required"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 128
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description",
                        "description": "Short description of the model, will be visible in the frontend.",
                        "update": "allowed"
                    },
                    "label": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 64
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Label",
                        "description": "Human friendly representation of the name/kind",
                        "update": "allowed"
                    },
                    "branch": {
                        "$ref": "#/components/schemas/BranchSupportType",
                        "description": "Type of branch support for the model.",
                        "default": "aware",
                        "update": "not_supported"
                    },
                    "default_filter": {
                        "anyOf": [
                            {
                                "type": "string",
                                "pattern": "^[a-z0-9\\_]*$"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Default Filter",
                        "description": "Default filter used to search for a node in addition to its ID. (deprecated: please use human_friendly_id instead)",
                        "update": "allowed"
                    },
                    "human_friendly_id": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Human Friendly Id",
                        "description": "Human friendly and unique identifier for the object.",
                        "update": "allowed"
                    },
                    "display_label": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Display Label",
                        "description": "Attribute or Jinja2 template to use to generate the display label",
                        "update": "allowed"
                    },
                    "display_labels": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string",
                                    "deprecationMessage": "display_labels are deprecated use display_label instead"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Display Labels",
                        "description": "List of attributes to use to generate the display label (deprecated)",
                        "update": "allowed"
                    },
                    "include_in_menu": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Include In Menu",
                        "description": "Defines if objects of this kind should be included in the menu.",
                        "update": "allowed"
                    },
                    "menu_placement": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Menu Placement",
                        "description": "Defines where in the menu this object should be placed.",
                        "update": "allowed"
                    },
                    "icon": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Icon",
                        "description": "Defines the icon to use in the menu. Must be a valid value from the MDI library https://icon-sets.iconify.design/mdi/",
                        "update": "allowed"
                    },
                    "order_by": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Order By",
                        "description": "List of entries to order results by. Supports attributes, relationship attributes, and node_metadata with __asc/__desc.",
                        "update": "allowed"
                    },
                    "uniqueness_constraints": {
                        "anyOf": [
                            {
                                "items": {
                                    "items": {
                                        "type": "string"
                                    },
                                    "type": "array"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Uniqueness Constraints",
                        "description": "List of multi-element uniqueness constraints that can combine relationships and attributes",
                        "update": "migration_required"
                    },
                    "documentation": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Documentation",
                        "description": "Link to a documentation associated with this object, can be internal or external.",
                        "update": "allowed"
                    },
                    "attributes": {
                        "items": {
                            "$ref": "#/components/schemas/AttributeSchema-Output"
                        },
                        "type": "array",
                        "title": "Attributes",
                        "description": "Node attributes",
                        "update": "not_applicable"
                    },
                    "relationships": {
                        "items": {
                            "$ref": "#/components/schemas/RelationshipSchema"
                        },
                        "type": "array",
                        "title": "Relationships",
                        "description": "Node Relationships",
                        "update": "not_applicable"
                    },
                    "inherit_from": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Inherit From",
                        "description": "List of Generic Kind that this profile is inheriting from"
                    },
                    "kind": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Kind"
                    },
                    "hash": {
                        "type": "string",
                        "title": "Hash"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "name",
                    "namespace",
                    "hash"
                ],
                "title": "APIProfileSchema"
            },
...
            "BranchSupportType": {
                "type": "string",
                "enum": [
                    "aware",
                    "agnostic",
                    "local"
                ],
                "title": "BranchSupportType"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "HashableModelState": {
                "type": "string",
                "enum": [
                    "present",
                    "absent"
                ],
                "title": "HashableModelState"
            },
...
            "SchemaNamespace": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "user_editable": {
                        "type": "boolean",
                        "title": "User Editable"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "user_editable"
                ],
                "title": "SchemaNamespace"
            },
            "SchemaReadAPI": {
                "properties": {
                    "main": {
                        "type": "string",
                        "title": "Main",
                        "description": "Main hash for the entire schema"
                    },
                    "nodes": {
                        "items": {
                            "$ref": "#/components/schemas/APINodeSchema"
                        },
                        "type": "array",
                        "title": "Nodes"
                    },
                    "generics": {
                        "items": {
                            "$ref": "#/components/schemas/APIGenericSchema"
                        },
                        "type": "array",
                        "title": "Generics"
                    },
                    "profiles": {
                        "items": {
                            "$ref": "#/components/schemas/APIProfileSchema"
                        },
                        "type": "array",
                        "title": "Profiles"
                    },
                    "templates": {
                        "items": {
                            "$ref": "#/components/schemas/APITemplateSchema"
                        },
                        "type": "array",
                        "title": "Templates"
                    },
                    "namespaces": {
                        "items": {
                            "$ref": "#/components/schemas/SchemaNamespace"
                        },
                        "type": "array",
                        "title": "Namespaces"
                    }
                },
                "type": "object",
                "required": [
                    "main"
                ],
                "title": "SchemaReadAPI"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            },
            "APIKeyHeader": {
                "type": "apiKey",
                "in": "header",
                "name": "X-INFRAHUB-KEY"
            }
        }
    }
}